chore(connect): restructure files in device folder
