+++ packages/connect/src/device/DeviceCommands.ts
@@ -3,9 +3,9 @@
 import { MessagesSchema as Messages } from '@trezor/protobuf';
 
 import { ERRORS, PROTO } from '../constants';
+import type { TypedCallProvider } from './DeviceCurrentSession';
 import { getBech32Network, getSegwitNetwork } from '../data/coinInfo';
-import type { TypedCallProvider } from '../device/DeviceCurrentSession';
-import { resolveDescriptorForTaproot } from '../device/resolveDescriptorForTaproot';
+import { resolveDescriptorForTaproot } from './utils/resolveDescriptorForTaproot';
 import type { HDNodeResponse } from '../types/api/getPublicKey';
 import type { BitcoinNetworkInfo, CoinInfo, Network } from '../types/coinInfo';
 import * as hdnodeUtils from '../utils/hdnodeUtils';
