+++ packages/connect/src/device/__tests__/resolveDescriptorForTaproot.test.ts
@@ -1,7 +1,7 @@
 import { MessagesSchema as Messages } from '@trezor/protobuf';
 
 import { HDNodeResponse } from '../../types/api/getPublicKey';
-import { resolveDescriptorForTaproot } from '../resolveDescriptorForTaproot';
+import { resolveDescriptorForTaproot } from '../utils/resolveDescriptorForTaproot';
 
 const originalResponse: HDNodeResponse = {
     path: [2147483734, 2147483648, 2147483648],
