+++ packages/connect/src/device/checks/__tests__/calculateRevisionForDevice.test.ts
