+++ packages/connect/src/device/checks/__tests__/checkFirmwareHashWithRetries.test.ts
@@ -1,10 +1,10 @@
 import type { Descriptor } from '@trezor/transport';
 import { Log } from '@trezor/utils';
 
-import type { DeviceUniquePath } from '../../types';
-import { Device } from '../Device';
-import * as checkFirmwareHashModule from '../workflow/checkFirmwareHash';
-import { checkFirmwareHashWithRetries } from '../workflow/checkFirmwareHashWithRetries';
+import type { DeviceUniquePath } from '../../../types';
+import { Device } from '../../Device';
+import * as checkFirmwareHashModule from '../checkFirmwareHash';
+import { checkFirmwareHashWithRetries } from '../checkFirmwareHashWithRetries';
 
 const { createTestTransport } = global.JestMocks;
