+++ packages/connect/src/device/checks/calculateRevisionForDevice.ts
