+++ packages/connect/src/device/checks/checkFirmwareHash.ts
