+++ packages/connect/src/device/checks/checkFirmwareHashWithRetries.ts
