+++ packages/connect/src/device/utils/resolveDescriptorForTaproot.ts
@@ -1,7 +1,7 @@
 import { MessagesSchema as Messages } from '@trezor/protobuf';
 import { convertTaprootXpub } from '@trezor/utils';
 
-import { HDNodeResponse } from '../types/api/getPublicKey';
+import { HDNodeResponse } from '../../types/api/getPublicKey';
 
 interface ResolveDescriptorForTaprootParams {
     response: HDNodeResponse;
